+++ schema/schema-update.sh
@@ -3,8 +3,8 @@
 THIS=$(dirname "$(readlink -f "$0")")
 image=maevsi/sqitch
 
-docker build -t "$image:build" --target build "$THIS/.." # --no-cache --progress plain
+sudo docker build -t "$image:build" --target build "$THIS/.." # --no-cache --progress plain
 
-container_id="$(docker create $image:build)"
-docker cp "$container_id:/srv/app/schema.sql" "$THIS/schema.definition.sql"
-docker rm -v "$container_id"
+container_id="$(sudo docker create $image:build)"
+sudo docker cp "$container_id:/srv/app/schema.sql" "$THIS/schema.definition.sql"
+sudo docker rm -v "$container_id"