+++ src/revert/enum_achievement_type.sql
@@ -1,0 +1,7 @@
+-- Revert maevsi:enum_achievement_type from pg
+
+BEGIN;
+
+DROP TYPE maevsi.achievement_type;
+
+COMMIT;