+++ src/revert/function_achievement_unlock.sql
@@ -1,0 +1,7 @@
+-- Revert maevsi:function_achievement_unlock from pg
+
+BEGIN;
+
+DROP FUNCTION maevsi.achievement_unlock;
+
+COMMIT;