+++ src/verify/function_account_password_reset.sql
@@ -4,7 +4,7 @@ BEGIN;
 
 DO $$
 BEGIN
-  ASSERT NOT (SELECT pg_catalog.has_function_privilege('maevsi_account', 'maevsi.account_password_reset(UUID, TEXT)', 'EXECUTE'));
+  ASSERT (SELECT pg_catalog.has_function_privilege('maevsi_account', 'maevsi.account_password_reset(UUID, TEXT)', 'EXECUTE'));
   ASSERT (SELECT pg_catalog.has_function_privilege('maevsi_anonymous', 'maevsi.account_password_reset(UUID, TEXT)', 'EXECUTE'));
 END $$;